pyproject.toml
[tool.poetry]
name = "sc2cast"
version = "0.1.0"
description = "Automated StarCraft II replay commentary and video generation"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "sc2cast", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
sc2reader = "^1.8.0"
pyautogui = "^0.9.54"
pywin32 = "^311"
pillow = "^12.0.0"
pytesseract = "^0.3.13"
easyocr = "^1.7.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
pytest-asyncio = "^0.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
